VMware: ensure rescue instance is deleted when...
instance is deleted

Backported fix created by Gary Kotton <gkotton@vmware.com>
https://review.openstack.org/#q,efb66531bc37ee416778a70d46c657608ca767af,n,z

If the user creates a rescue instance and then proceeded to delete
the original instance then the rescue instance would still be up
and running on the backend.

This patch ensures that the rescue instance is cleaned up if
necessary.

The vmops unrescue method has a new parameter indicating if
the original VM should be powered on.